Keep all json structs in json_span.go
