--- data.go
@@ -1,6 +1,0 @@
-package instana
-
-type Data struct {
-	Service string   `json:"service,omitempty"`
-	SDK     *SDKData `json:"sdk"`
-}
